chore: reorganize routing internals
