+++ src/application/create-application.ts
@@ -5,7 +5,7 @@ import { applyConfiguredGlobalMiddleware } from '@/Http/middleware/apply-configu
 import { initializeRequestScopeStorage } from '@/Http/Scope/request-scope-storage';
 import { registerEventSubscribers } from '@/Kernel';
 import { registerLegacyRoutes } from '@/routing/decorator/legacy-router';
-import { registerRoutes } from '@/routing/register-routes';
+import { registerRoutes } from '@/routing/registration/register-routes';
 import { verifyRoutingMode } from '@/routing/verify-routing-mode';
 import Koa from 'koa';
 import { type Application } from './application';
