+++ src/routing/definition/resolve-route-options.test.ts
