+++ src/routing/index.ts
@@ -1,8 +1,8 @@
 export { Route } from './route';
-export { RouteGroup } from './route-group';
-export { Any, Delete, Get, Head, Options, Patch, Post, Put } from './http-verb-helpers';
+export { RouteGroup } from './helpers/route-group';
+export { Any, Delete, Get, Head, Options, Patch, Post, Put } from './helpers/http-verb-helpers';
 export type { RouteDeclaration } from './route';
-export type { RouteConfigOverlay, RouteGroupDefinition, RouteGroupOptions } from './route-group';
-export type { RouteSource } from './route-source';
+export type { RouteConfigOverlay, RouteGroupDefinition, RouteGroupOptions } from './helpers/route-group';
+export type { RouteSource } from './source/route-source';
 export type { HttpMethod } from './http-method';
 export type { RouteOptions } from './route-options';
