+++ src/routing/registration/expand-route-definitions.test.ts
