+++ src/routing/registration/expand-route-definitions.ts
@@ -1,6 +1,6 @@
 import { koaBody } from 'koa-body';
-import type { RouteDefinition } from '@/routing/route-definition';
-import type { RouteRegistration } from '@/routing/route-registration';
+import type { RouteDefinition } from '@/routing/definition/route-definition';
+import type { RouteRegistration } from '@/routing/registration/route-registration';
 
 export function expandRouteDefinitions(routes: RouteDefinition[]): RouteRegistration[] {
   const registrations: RouteRegistration[] = [];
