+++ src/routing/registration/register-routes.test.ts
@@ -2,10 +2,10 @@ import Koa from 'koa';
 import supertest from 'supertest';
 import { describe, expect, test } from 'vitest';
 import { type Application } from '@/application/application';
-import { Get } from './http-verb-helpers';
+import { Get } from '@/routing/helpers/http-verb-helpers';
 import { registerRoutes } from './register-routes';
-import { RouteGroup } from './route-group';
-import { Route } from './route';
+import { RouteGroup } from '@/routing/helpers/route-group';
+import { Route } from '@/routing/route';
 
 describe('register routes', () => {
   test('it registers a function-first route', async () => {
