--- src/routing/route-source.ts
@@ -1,4 +1,0 @@
-import type { RouteDefinition } from '@/routing/route-definition';
-import type { RouteGroupDefinition } from '@/routing/route-group';
-
-export type RouteSource = RouteDefinition | RouteGroupDefinition;
