+++ src/routing/source/route-source.ts
@@ -1,0 +1,4 @@
+import type { RouteDefinition } from '@/routing/definition/route-definition';
+import type { RouteGroupDefinition } from '@/routing/helpers/route-group';
+
+export type RouteSource = RouteDefinition | RouteGroupDefinition;
